Fix publishing latest to staging
TDS used to have a `dev` and a `main` branch. Latest staging packages were built off of `dev` and official releases were in `main`. When we switch to having `main` be the latest release and version tags be the official production release, the `dev` artifact was left in the publish action.