Fix es client so that it does not get created prior to api kick-off
This fixes the bug that causes failures when trying to use a ES client
from inside a child process that is defined in the parent process. This
causes errors and failures when communicating with ES.

The fix chosen was to not spawn a global es client at module import time
and instead to provide a factory function for fetching a new client
when needed.